+++ changelog.md
@@ -1,6 +1,7 @@
 * unreleased
     - Drop support for GHC < 8.4, require base > 4.11.
     - Bump `cabal-version` to 2.4.
+    - Add `Ord` instance for `PlaylistName`, `Path`, and `Value`
     
 * v0.9.2.0 2020-10-02
     - New command: `seekCur`